+++ mission/landmark_server/CMakeLists.txt
@@ -29,6 +29,7 @@ set(LIB_NAME "${PROJECT_NAME}_component")
 
 add_library(${LIB_NAME} SHARED
   src/landmark_server_ros.cpp
+  src/landmark_server_convergence.cpp
   src/landmark_server_ros_debug.cpp
   src/landmark_ros_conversion.cpp)
