+++ mission/landmark_server/config/landmark_server_config.yaml
@@ -29,8 +29,6 @@
         prob_of_survival: 0.95
         estimate_clutter: false
 
-    body_z_offset: -0.233 # meter Constant z offset applied to reference filter target pose
-
     debug:
       enable: false
       topic_name: "landmark_server/debug/landmark_tracks"
